Revert Origin's Equality_comparable and Weakly_ordered concept requirements to avoid use of constrained-type-specifers in the trailing-return-type of compound-requirements.
Commit 8edde4176f3ce0e85c95b1e6b7b006c2ef5a6d7e changed Origin's
definitions of the Equality_comparable and Weakly_ordered concepts
to specify requirements in terms of compound-requirements that use
constrained-type-specifers in their trailing-return-type instead of
non-constrained simple type specifiers.  This change fell afoul of
an apparent gcc defect that results in spurious ambiguous template
instantiation errors.  The gcc defect is tracked at [1].  This
commit reverts that change as a workaround for the gcc defect.

[1]: https://gcc.gnu.org/bugzilla/show_bug.cgi?id=69235