Revert Origin's Integral_type concept to use std::is_integral.
Commit 75cedb4debb015648477926393c668c7788263ae changed Origin's
definition of the Integral_type in terms of Origin provided core
traits classes (later changed to constexpr functions) instead of
std::is_integral.  This change ran afoul of an apparent gcc defect
that results in catastrophic compile time performance impact for
uses of this concept in some contexts.  The gcc defect is tracked
at [1].  This commit reverts that change as a workaround for the
gcc defect.

[1]: https://gcc.gnu.org/bugzilla/show_bug.cgi?id=67565